Moving the starting hex selection code.
Taking this out of the robot class and putting into the world class.
The world class will tell the robot where it needs to start from and we
will inject that into the robot initializer.

Also, need to remove the "selected" property of the hexagon and add in a
contains of different base object types so that we can have other types
of objects inhabit the world.