The bot doesn't have a current hex when it's created.
It gets a current hex when it is added to a world object.
Also need to move all the world scanning code out of the bot and into
the world.  The bot scans the world and the world returns a list of
adjacent sqaure.  The bot can then decide if it can move to them based
on height/distance.

We might need to allow the bot to decide an impossible move, say "Move
to NE to hex with height of 8" and the bot is at 4.  This move would not
be allowed and the world would return the result.